Add files via upload
The dataset was collected from Airbnb containing details such as listing amenities, review ratings, neighbourhoods, and availability, with a total of 75 columns. 
The first part of the task worked in Jupyter, was exploratory, exploring trends in the dataset using data visualization techniques such as boxplots, scatter plots, and histograms. 
Additionally, some statistical analysis on trends was done.
For the second part of the task worked in VS Code, a command-line interface was created. through which users can interact with the dataset by entering queries in the terminal. That will enable users to retrieve specific information, based on their queries.